🏗️: keep DTOs in objects module
